backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
requests
pydantic
passlib[bcrypt]
python-jose[cryptography]
google-generativeai
yfinance
duckduckgo-search
pandas
google-auth
python-multipart
